+++ ktor-client/ktor-client-winhttp/windows/src/io/ktor/client/engine/winhttp/internal/WinHttpWebSocket.kt
@@ -229,7 +229,7 @@ internal class WinHttpWebSocket(
                     hWebSocket,
                     status.ptr,
                     null,
-                    0,
+                    0.toUInt(),
                     reasonLengthConsumed.ptr
                 ) != 0u
             ) {
@@ -273,7 +273,7 @@ internal class WinHttpWebSocket(
             hWebSocket,
             WINHTTP_WEB_SOCKET_SUCCESS_CLOSE_STATUS.convert(),
             NULL,
-            0
+            0.toUInt()
         )
         platform.winhttp.WinHttpCloseHandle(hWebSocket)
         connect.close()
